fixed take and add anything bug AB#773, navigate to amalfi bug AB#771, refactored to seperate functionality from PizzaQuestApp AB#510
